requirements.txt
docling>=2.0.0
Pillow>=10.0.0
python-pptx>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=0.3.0
faiss-cpu>=1.7.0
pydantic>=2.0.0
